+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v4.0.6
 ### Changes
+* Remove `Profile Time Used` from the optimizer settings as the log file now always contains the profiling time.
+* Add global setting `Profile Time Used in UI` to show the profiling time for the editor UI at the bottom of the inspector. 
 * Add more granular logging of performance of different parts of the optimizer to the log file.
 * Rename `Material Merge Analyzer` button to `Open Material Merge Analyzer`.